generalize auth function out of events app
