+++ website/events/helpers/tests.py
@@ -1,7 +1,7 @@
 from django.contrib.auth.models import Permission
 from django.contrib.contenttypes.models import ContentType
 
-from events.constants import CAN_VIEW_ORGANIZERS_CODENAME
+from events.constants import CAN_VIEW_EVENT_ORGANIZERS_CODENAME, CAN_VIEW_ORGANIZERS_CODENAME
 from events.models import Event, Organizer, EventOrganizer, SponsorCategory
 
 def organizer_permissions():
@@ -26,13 +26,15 @@ def super_organizer_permissions():
     
     permissions.append(Permission.objects.get(content_type=event_content_type, codename='add_event'))
     # Event create permmission.
-    permissions.append(Permission.objects.get(content_type=event_content_type, codename=CAN_VIEW_ORGANIZERS_CODENAME))
+    permissions.append(Permission.objects.get(content_type=event_content_type, codename=CAN_VIEW_EVENT_ORGANIZERS_CODENAME))
     # Can view associated event organizers 
 
     permissions.append(Permission.objects.get(content_type=organizer_content_type, codename='add_organizer'))
     # Organizer create permmission.
     permissions.append(Permission.objects.get(content_type=organizer_content_type, codename='change_organizer'))
     # Organizer change permmission.
+    permissions.append(Permission.objects.get(content_type=organizer_content_type, codename=CAN_VIEW_ORGANIZERS_CODENAME))
+    # Can view organizers permmission. 
 
     permissions.append(Permission.objects.get(content_type=organizer_content_type, codename='add_organizer'))
     # Organizer add permmission.
