+++ website/events/static/css/events.css
@@ -2,23 +2,6 @@
     text-align: center;
 }
 
-.events-login-form {
-    width: 340px;
-    margin: 50px auto;
-}
-.events-login-form .asteriskField {
-    display: none;
-}
-
-.events-login-form form {
-    margin-bottom: 15px;
-    background: #f7f7f7;
-    box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
-    padding: 30px;
-}
-.events-login-form h2 {
-    margin: 0 0 15px;
-}
 .events-form-control, .btn {
     min-height: 38px;
     border-radius: 2px;
