+++ website/members/templates/base.html
@@ -21,8 +21,9 @@
 		<link rel="stylesheet" href="{% static 'css/bootstrap.min.css' %}">
 		<!-- font-awesome -->
 		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+		<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.8.1/css/all.css" integrity="sha384-50oBUHEmvpQ+1lW4y57PTFmhCaXp0ML5d60M1M7uH2+nqUivzIebhndOJK28anvf" crossorigin="anonymous">
 		<!-- google font -->
-	    <link href='http://fonts.googleapis.com/css?family=Open+Sans:400,300,400italic,700,800' rel='stylesheet' type='text/css'>
+		<link href='http://fonts.googleapis.com/css?family=Open+Sans:400,300,400italic,700,800' rel='stylesheet' type='text/css'>
 
 		<!-- custom css -->
 		<link rel="stylesheet" href="{% static 'css/templatemo-style.css' %}">
@@ -51,7 +52,23 @@
 						<!-- <li><a href="#feature">Misión</a></li> -->
 						<li><a href="https://ac.python.org.ar">Asociación Civil</a></li>
 					</ul>
-					{% block extra_navbar_menu %}{% endblock %}
+					{% block extra_navbar_menu %}
+					{% endblock %}
+
+					{% if user.is_authenticated %}
+					<ul class="nav navbar-nav navbar-right">
+						<li class="nav-item dropdown">
+							<a class="nav-link dropdown-toggle" id="navbarDropdownMenuLink-4" data-toggle="dropdown"
+							aria-haspopup="true" aria-expanded="false" href="#">
+							<i class="fas fa-user"></i> {{ user.username }} </a>
+							<ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink-4">
+								<li><a href="{% url 'profile' %}">Datos cuenta</a></li>
+								<li><a href="{% url 'logout' %}">Log out</a></li>
+							</ul>
+						</li>
+					</ul>
+					{% endif %}
+
 				</div>
 			</div>
 		</nav>
