+++ website/static/css/templatemo-style.css
@@ -588,3 +588,21 @@ ul.fa-ul li {
   margin: 0 0 20px 20px;
   padding: 12px 0 0 20px;
 }
+
+.login-form {
+    width: 340px;
+    margin: 50px auto;
+}
+.login-form .asteriskField {
+    display: none;
+}
+
+.login-form form {
+    margin-bottom: 15px;
+    background: #f7f7f7;
+    box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.3);
+    padding: 30px;
+}
+.login-form h2 {
+    margin: 0 0 15px;
+}
