+++ website/templates/registration/change_password.html
@@ -1,4 +1,4 @@
-{% extends 'events_base.html' %}
+{% extends 'base.html' %}
 {% load crispy_forms_tags %}
 {% block title %}Cambio de Contraseña{% endblock %}
