+++ website/templates/registration/logged_out.html
@@ -1,4 +1,4 @@
-{% extends "events_base.html" %}
+{% extends "base.html" %}
 
 {% block content %}
 <div class="container">
