+++ website/templates/registration/login.html
@@ -1,4 +1,4 @@
-{% extends "events_base.html" %}
+{% extends "base.html" %}
 {% load crispy_forms_tags %}
 {% block content %}
 <div class="container">
@@ -17,7 +17,7 @@
     {% endif %}
 {% endif %}
 
-<div class="events-login-form">
+<div class="login-form">
     <form method="post" action="{% url 'login' %}">
     <h2 class="text-center">Inicio de sesión</h2> 
     {% csrf_token %}
