+++ website/templates/registration/password_reset_complete.html
@@ -1,4 +1,4 @@
-{% extends 'events_base.html' %}
+{% extends 'base.html' %}
 
 {% block content %}
 <div class="container">
