+++ website/templates/registration/password_reset_confirm.html
@@ -1,4 +1,4 @@
-{% extends 'events_base.html' %}
+{% extends 'base.html' %}
 {% load crispy_forms_tags %}
 {% block content %}
 <div class="container">
