+++ website/templates/registration/password_reset_done.html
@@ -1,5 +1,4 @@
-<!-- templates/registration/password_reset_done.html -->
-{% extends 'events_base.html' %}
+{% extends 'base.html' %}
 
 {% block title %}Correo enviado{% endblock %}
