+++ website/templates/registration/password_reset_form.html
@@ -1,5 +1,4 @@
-<!-- templates/registration/password_reset_form.html -->
-{% extends 'events_base.html' %}
+{% extends 'base.html' %}
 {% load crispy_forms_tags %}
 {% block title %}Reseteo de Contraseña{% endblock %}
